% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotplot_module.R
\name{dotplot_server}
\alias{dotplot_server}
\title{Produces Server for Gene Dotplot Module}
\usage{
dotplot_server(id)
}
\arguments{
\item{id}{module id}
}
\description{
Produces Server for Gene Dotplot Module
}
