% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_fill_hue.R
\name{gg_fill_hue}
\alias{gg_fill_hue}
\title{Generate color palettes}
\usage{
gg_fill_hue(n)
}
\arguments{
\item{n}{the number of colors to generate}
}
\value{
a color for each n
}
\description{
Generate color palettes
}
