% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gene_taxa_RAs.R
\name{plot_gene_taxa_RAs}
\alias{plot_gene_taxa_RAs}
\title{Plot Gene Counts by Taxonomic Groups}
\usage{
plot_gene_taxa_RAs(metagenome_data, genes, taxa_level, font_size, cols, border)
}
\arguments{
\item{metagenome_data}{metagenomic data in long form with columns for genes counts, orthology groups,
sample data and taxonomic data}

\item{genes}{A string. The gene to plot}

\item{taxa_level}{A string, the taxonomic level to group by for plotting}

\item{font_size}{An integer. Plot font size}

\item{cols}{Colors for taxonomic groups, a character vector of length n, where n is the number of taxonomic levels}

\item{border}{An integer. Determines plot border thickness}
}
\value{
A ggplot2 object
}
\description{
Plot Gene Counts by Taxonomic Groups
}
