% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RA_plot_module.R
\name{RA_server}
\alias{RA_server}
\title{Produces the Server for the Gene Counts by Taxa Module}
\usage{
RA_server(id)
}
\arguments{
\item{id}{module id}
}
\description{
Produces the Server for the Gene Counts by Taxa Module
}
