% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RA_plot_module.R
\name{RA_timepoint_color}
\alias{RA_timepoint_color}
\title{Produces Color Palette Widget}
\usage{
RA_timepoint_color(id)
}
\arguments{
\item{id}{module id}
}
\description{
Produces Color Palette Widget
}
