% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RA_plot_module.R
\name{RA_ui}
\alias{RA_ui}
\title{Produces the UI for the Gene Counts by Taxa Module}
\usage{
RA_ui(id)
}
\arguments{
\item{id}{module id}
}
\description{
Produces the UI for the Gene Counts by Taxa Module
}
