% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcano_module.R
\name{volcano_server}
\alias{volcano_server}
\title{Produces Server for Volcano Plot Module}
\usage{
volcano_server(id)
}
\arguments{
\item{id}{module id}
}
\description{
Produces Server for Volcano Plot Module
}
